Scope volunteer shifts by year
Extract a common mixin for year-scoping
