+++ app/models/concerns/year_scoped.rb
@@ -1,0 +1,23 @@
+module YearScoped
+  extend ActiveSupport::Concern
+
+  included do
+    after_initialize do
+      self.year ||= Date.today.year
+    end
+  end
+
+  class_methods do
+    def for_current_year
+      where(year: Date.today.year)
+    end
+
+    def for_previous_years
+      where('year < ? ', Date.today.year)
+    end
+  end
+
+  def for_current_year?
+    year == Date.today.year
+  end
+end
