+++ db/migrate/20160804150055_add_year_to_volunteer_shifts.rb
@@ -1,0 +1,5 @@
+class AddYearToVolunteerShifts < ActiveRecord::Migration
+  def change
+    add_column :volunteer_shifts, :year, :integer
+  end
+end
