+++ spec/models/volunteership_spec.rb
@@ -7,5 +7,4 @@ RSpec.describe Volunteership, type: :model do
   it 'defaults its year to the current year' do
     expect(Volunteership.new.year).to eq(Date.today.year)
   end
-
 end
